+++ ArcaneWizardLibrary.lua
@@ -2,6 +2,7 @@ local addonName, LIB = ...
 
 ArcaneWizardLibrary = ArcaneWizardLibrary or {}
 ArcaneWizardLibrary.Dialogs = {}
+ArcaneWizardLibrary.Profiles = {}
 ArcaneWizardLibrary.Settings = {}
 
 ArcaneWizardLibrary.ADDON_AUTHOR = C_AddOns.GetAddOnMetadata(addonName, "Author")