+++ CHANGELOG.md
@@ -2,6 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Library/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Library/releases)
 
-- Added: Optional cancel callback support for confirmation dialogs
-- Added: TOC version for patch 12.0.7 [retail]
-
+- Added: Profile helper API for character keys and localized profile texts